refactor(chalice): refactored errors helper refactor(chalice): removed errors-tags
